chore: Refactoring VertexAiSessionServiceTest's mock ApiAnswer method to facilitate future changes.
The goal is to fix appendEvent_success().  In order to do that, I'll need to do some non trivial changes to the `Answer<ApiResponse>` in the setup.

The refactor mostly just breaks out the logic to handle different URL patterns into separate methods.  This will make future changes related to Event handling easier to understand.

PiperOrigin-RevId: 767697219